Delete old Cursor; rename Cursor2 → Cursor
The single-buffer Cursor (formerly `Cursor2`) is now the only one.
- `lib/zephyrine/src/core/zephyrine.Cursor.scala` — the old block/Iterator-based implementation, deleted; replaced by what was `Cursor2.scala`.
- `Cursor2` → `Cursor`, `Cursor2.Mark` → `Cursor.Mark`, `Cursor2.Held` → `Cursor.Held` everywhere.
- `internal.consume2` collapses back into `internal.consume`.
- Tests/benches drop the duplicated `Cursor2` mirrors and the now-meaningless `blockTail` / `advanceBlock` tests (the new model has no notion of blocks).

Tests: zephyrine 189/189, merino 317/317, honeycomb 296/296, xylophone 140/140, gesticulate 57/57, telekinesis 110/13 (matches main), obligatory 12/12 (via assembly; the test-runner package mismatch is pre-existing). zeppelin compiles.
