+++ lib/obligatory/src/core/obligatory.CarriageReturn.scala
@@ -41,7 +41,7 @@ import zephyrine.*
 
 object CarriageReturn:
   given framable: Text is Framable by CarriageReturn = input =>
-    val cursor = Cursor2(input)
+    val cursor = Cursor(input)
 
     def frame(): Optional[Text] = cursor.hold:
       val start = cursor.mark
