+++ lib/obligatory/src/core/obligatory.LengthPrefix.scala
@@ -43,7 +43,7 @@ object LengthPrefix:
   given framable: Tactic[FrameError] => Data is Framable by LengthPrefix = input =>
     def fail() = abort(FrameError(FrameError.Reason.ShortRead))
 
-    val cursor = Cursor2(input)
+    val cursor = Cursor(input)
 
     def length: Optional[Int] =
       cursor.lay(Unset): byte0 =>
