+++ lib/zephyrine/src/core/zephyrine.internal.scala
@@ -57,24 +57,3 @@ object internal:
         recur(index + 1, checks2)
 
     recur(0, '{()})
-
-  def consume2(cursor: Expr[Cursor2[?]], text0: Expr[String], otherwise: Expr[Unit]): Macro[Unit] =
-    import quotes.reflect.*
-
-    val text = text0.valueOrAbort
-
-    def recur(index: Int, checks: Expr[Unit]): Expr[Unit] =
-      if index >= text.length then checks else
-        val char = text(index)
-
-        val checks2 =
-          ' {
-              $checks
-              $cursor.next()
-              $cursor.lay($otherwise): datum =>
-                if datum != ${Expr(char)} then $otherwise
-            }
-
-        recur(index + 1, checks2)
-
-    recur(0, '{()})
